refactor: simplify advanced operation handling and improve iptables chain info loading
